refactor: clippy fixes + channels default + cleanup
Fix 39 clippy warnings (dead_code, unused_imports, await_holding_lock, needless_lifetimes, result_large_err, collapsible_match, etc.)

Enable channels feature by default in Cargo.toml (P1-B-2)

Remove unused duplicate module files (bootstrap.rs/config.rs/headless.rs/setup.rs) created by prior agent - kept lib.rs stable

Update StreamEvents doc comment (streaming already implemented)

Add temp files to .gitignore (count_panic.py, build-output.txt, etc.)

Verified: cargo check + cargo clippy --features grpc,channels -- -D warnings pass

+++ src-tauri/src/memory/sqlite_store.rs
@@ -768,7 +768,7 @@ impl SqliteStore {
                         params![id],
                     )
                     .map_err(|e| anyhow!("sqlite archive_memories error: {e}"))?;
-                total += n as usize;
+                total += n;
             }
             Ok(total)
         })
